+++ client/src/pages/Account/AccountNotifications/AccountNotifications.tsx
@@ -1,0 +1,5 @@
+import { JSX } from 'react';
+
+export default function AccountNotifications(): JSX.Element {
+  return <></>;
+}