+++ client/src/pages/Account/contexts/AccountLocationContext.ts
@@ -1,0 +1,11 @@
+import { createContext, Dispatch, SetStateAction } from 'react';
+
+export type AccountLocation = 'profile' | 'social' | 'notifications' | 'wishlists';
+
+export type AccountLocationContextType = {
+  accountLocation: AccountLocation;
+  setAccountLocation: Dispatch<SetStateAction<AccountLocation>>;
+};
+
+const AccountLocationContext = createContext<AccountLocationContextType | null>(null);
+export default AccountLocationContext;