+++ src/db/helpers/accountDbHelpers.ts
@@ -1,6 +1,9 @@
-import { Request } from 'express';
+import { Request, Response } from 'express';
 import { Pool, PoolConnection, ResultSetHeader } from 'mysql2/promise';
 import { logUnexpectedError } from '../../logs/errorLogger';
+import { ACCOUNT_FAILED_SIGN_IN_LIMIT } from '../../util/constants/accountConstants';
+import { removeRequestCookie } from '../../util/cookieUtils';
+import { purgeAuthSessions } from '../../auth/authSessions';
 
 export async function deleteAccountById(accountId: number, executor: Pool | PoolConnection, req: Request): Promise<boolean> {
   try {
@@ -22,7 +25,7 @@ export async function deleteAccountById(accountId: number, executor: Pool | Pool
 }
 
 export async function incrementVerificationEmailsSent(
-  verificationId: Readonly<number>,
+  verificationId: number,
   executor: Pool | PoolConnection,
   req: Request
 ): Promise<boolean> {
@@ -71,6 +74,56 @@ export async function incrementFailedVerificationAttempts(
   }
 }
 
+export async function incrementEmailChangeEmailsSent(
+  emailUpdateId: number,
+  executor: Pool | PoolConnection,
+  req: Request
+): Promise<boolean> {
+  try {
+    const [resultSetHeader] = await executor.execute<ResultSetHeader>(
+      `UPDATE
+        email_update
+      SET
+        update_emails_sent = update_emails_sent + 1
+      WHERE
+        update_id = ?;`,
+      [emailUpdateId]
+    );
+
+    return resultSetHeader.affectedRows > 0;
+  } catch (err: unknown) {
+    console.log(err);
+    await logUnexpectedError(req, err, 'failed to increment update_emails_sent');
+
+    return false;
+  }
+}
+
+export async function incrementedFailedEmailChangeAttempts(
+  verificationId: number,
+  executor: Pool | PoolConnection,
+  req: Request
+): Promise<boolean> {
+  try {
+    const [resultSetHeader] = await executor.execute<ResultSetHeader>(
+      `UPDATE
+        email_update
+      SET
+        failed_update_attempts = failed_update_attempts + 1
+      WHERE
+        update_id = ?;`,
+      [verificationId]
+    );
+
+    return resultSetHeader.affectedRows > 0;
+  } catch (err: unknown) {
+    console.log(err);
+    await logUnexpectedError(req, err, 'failed to increment failed_update_attempts');
+
+    return false;
+  }
+}
+
 export async function incrementFailedSignInAttempts(accountId: number, executor: Pool | PoolConnection, req: Request): Promise<boolean> {
   try {
     const [resultSetHeader] = await executor.execute<ResultSetHeader>(
@@ -111,4 +164,25 @@ export async function resetFailedSignInAttempts(accountId: number, executor: Poo
 
     return false;
   }
+}
+
+export async function handleIncorrectPassword(
+  accountId: number,
+  failedSignInAttempts: number,
+  executor: Pool | PoolConnection,
+  req: Request,
+  res: Response
+): Promise<void> {
+  const incremented: boolean = await incrementFailedSignInAttempts(accountId, executor, req);
+  const hasBeenLocked: boolean = failedSignInAttempts + 1 >= ACCOUNT_FAILED_SIGN_IN_LIMIT && incremented;
+
+  if (hasBeenLocked) {
+    removeRequestCookie(res, 'authSessionId');
+    await purgeAuthSessions(accountId);
+  }
+
+  res.status(401).json({
+    message: `Incorrect password.${hasBeenLocked ? ' Account locked.' : ''}`,
+    reason: hasBeenLocked ? 'incorrectPassword_locked' : 'incorrectPassword',
+  });
 }