+++ src/util/email/emailTemplates/accountVerificationEmailTemplate.ts
@@ -1,7 +1,7 @@
 import { minuteMilliseconds } from '../../constants/globalConstants';
 import { ACCOUNT_VERIFICATION_WINDOW } from '../../constants/accountConstants';
 
-type AccountVerificationEmailTemplate = {
+export type AccountVerificationEmailTemplate = {
   displayName: string;
   publicAccountId: string;
   verificationToken: string;
@@ -23,7 +23,6 @@ export default function accountVerificationEmailTemplate({
           name="viewport"
           content="width=device-width, initial-scale=1.0"
         />
-        <title>Email</title>
       </head>
       
       <body>
@@ -38,9 +37,9 @@ export default function accountVerificationEmailTemplate({
             >.
           </p>
 
-          <p>Your account will be automatically deleted if it's not verified within ${
+          <p>Your account will be automatically deleted if it's not verified within ${Math.floor(
             ACCOUNT_VERIFICATION_WINDOW / minuteMilliseconds
-          } minutes of being created.</p>
+          )} minutes of being created.</p>
           <p>If this request wasn't made by you, feel free to ignore it.</p>
 
           <p">Warmest regards,</p>